Added signature validation to test which fails
